bugfix time integration

+++ src/profile.h
@@ -7,4 +7,4 @@ void Profile_Info(const char* file, const char* func, const int line,
 void Profile_Report();
 void Profile_Report_Last();
 void Write_Logs();
-float Runtime();
+double Runtime();
